# api-docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ADP 鱼塘养殖企业 API",
    "version": "1.0.0",
    "description": "网站二正式业务接口。提交后可修改并递增版本，核验后只读。"
  },
  "servers": [
    {
      "url": "https://1.14.148.15",
      "description": "生产环境"
    }
  ],
...
  "paths": {
...
    "/api/v1/master-data/{resource}": {
      "post": {
        "tags": [
          "master-data"
        ],
        "summary": "新建记录",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "master_data_create_record_post_api_v1_master_data_resource",
        "parameters": [
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      },
      "get": {
        "tags": [
          "master-data"
        ],
        "summary": "查询详情",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "master_data_list_records_get_api_v1_master_data_resource",
        "parameters": [
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "CookieAuth": []
          }
        ]
      }
    },
    "/api/v1/master-data/{resource}/{record_id}": {
      "delete": {
        "tags": [
          "master-data"
        ],
        "summary": "删除未提交草稿",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "master_data_delete_record_delete_api_v1_master_data_resource_record_id",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      },
      "get": {
        "tags": [
          "master-data"
        ],
        "summary": "查询详情",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "master_data_get_record_get_api_v1_master_data_resource_record_id",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "security": [
          {
            "CookieAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "master-data"
        ],
        "summary": "更新",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "master_data_update_record_patch_api_v1_master_data_resource_record_id",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/v1/master-data/{resource}/{record_id}/archive": {
      "post": {
        "tags": [
          "master-data"
        ],
        "summary": "新建记录",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "master_data_archive_record_post_api_v1_master_data_resource_record_id_archive",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/v1/master-data/{resource}/{record_id}/submit": {
      "post": {
        "tags": [
          "master-data"
        ],
        "summary": "提交核验",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "master_data_submit_record_post_api_v1_master_data_resource_record_id_submit",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/v1/master-data/{resource}/{record_id}/verify": {
      "post": {
        "tags": [
          "master-data"
        ],
        "summary": "核验",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "master_data_verify_record_post_api_v1_master_data_resource_record_id_verify",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "resource",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/v1/master-data/ponds/{pond_id}/status-changes": {
      "post": {
        "tags": [
          "master-data"
        ],
        "summary": "新建记录",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "master_data_request_pond_status_change_post_api_v1_master_data_ponds_pond_id_status_changes",
        "parameters": [
          {
            "name": "pond_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/v1/master-data/ponds/{pond_id}/status-changes/{request_id}/verify": {
      "post": {
        "tags": [
          "master-data"
        ],
        "summary": "核验",
        "description": "核验完成的数据只读；DELETE 仅适用于从未提交且无业务引用的草稿。",
        "operationId": "master_data_verify_pond_status_change_post_api_v1_master_data_ponds_pond_id_status_changes_request_id_verify",
        "parameters": [
          {
            "name": "pond_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutationRequest"
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": [],
            "CsrfToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "adp_session"
      },
      "CsrfSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      },
      "CsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "OK"
          },
          "message": {
            "type": "string",
            "example": "操作成功"
          },
          "data": {},
          "request_id": {
            "type": "string",
            "example": "request-example-001"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "string",
            "example": "VERSION_CONFLICT"
          },
          "message": {
            "type": "string",
            "example": "操作成功"
          },
          "data": {},
          "request_id": {
            "type": "string",
            "example": "request-example-001"
          }
        }
      },
      "MutationRequest": {
        "type": "object",
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1
          },
          "evidence_attachment_ids": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          }
        },
        "additionalProperties": true
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "统一错误；以 code 和 request_id 定位问题",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}